+++ coursemos_spring_boot/src/main/java/com/example/demo/service/CompanyServiceImpl.java
@@ -31,7 +31,7 @@ public class CompanyServiceImpl implements CompanyService {
 	
 
 	@Override
-	public Company getCompanyByPlaceId(int placeId) {
+	public Company getCompanyByPlaceId(Long placeId) {
 		return companyDao.getCompanyByPlaceId(placeId);
 	}
